backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
python-docx
PyPDF2
pydantic
python-multipart
numpy
google-generativeai
python-dotenv
aiofiles
requests
